requirement.txt
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-rest-framework==0.1.0
djangorestframework-jwt==1.11.0
pillow==6.0.0